Update shared kernel structure
