+++ src/Urleso.Application/Abstractions/Services/ICodeGeneratingService.cs
@@ -1,4 +1,4 @@
-﻿using Urleso.SharedKernel;
+﻿using Urleso.SharedKernel.Results;
 
 namespace Urleso.Application.Abstractions.Services;
