+++ src/Urleso.Application/ShortenedUrls/CreateShortenedUrl/CreateShortenedUrlCommandHandler.cs
@@ -3,7 +3,7 @@ using Urleso.Application.Abstractions.Data.Repositories;
 using Urleso.Application.Abstractions.Messaging;
 using Urleso.Application.Abstractions.Services;
 using Urleso.Domain.ShortenedUrls;
-using Urleso.SharedKernel;
+using Urleso.SharedKernel.Results;
 
 namespace Urleso.Application.ShortenedUrls.CreateShortenedUrl;
