+++ src/Urleso.SharedKernel/Results/NotAvailableErrorException.cs
@@ -1,4 +1,4 @@
-namespace Urleso.SharedKernel;
+namespace Urleso.SharedKernel.Results;
 
 public sealed class NotAvailableErrorException() : Exception(DetailsMessage)
 {
