+++ src/Urleso.SharedKernel/Results/Result.cs
@@ -1,4 +1,4 @@
-﻿namespace Urleso.SharedKernel;
+﻿namespace Urleso.SharedKernel.Results;
 
 public sealed class Result
 {
