--- tests/Urleso.SharedKernel.UnitTests/DummyEntity.cs
@@ -1,3 +1,0 @@
-﻿namespace Urleso.SharedKernel.UnitTests;
-
-internal sealed class DummyEntity(DummyEntityId id) : Entity<DummyEntityId>(id);
