--- tests/Urleso.SharedKernel.UnitTests/DummyEntityId.cs
@@ -1,3 +1,0 @@
-﻿namespace Urleso.SharedKernel.UnitTests;
-
-internal sealed record DummyEntityId(Guid Value);
