+++ tests/Urleso.SharedKernel.UnitTests/Entities/DummyEntity.cs
@@ -1,0 +1,5 @@
+﻿using Urleso.SharedKernel.Entities;
+
+namespace Urleso.SharedKernel.UnitTests.Entities;
+
+internal sealed class DummyEntity(DummyEntityId id) : Entity<DummyEntityId>(id);
