+++ tests/Urleso.SharedKernel.UnitTests/Entities/DummyEntityId.cs
@@ -1,0 +1,3 @@
+﻿namespace Urleso.SharedKernel.UnitTests.Entities;
+
+internal sealed record DummyEntityId(Guid Value);
